fixing fitur 9

+++ siruangan/src/main/java/com/apap/tugasakhir/siruangan/repository/PeminjamanRuanganDB.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PeminjamanRuanganDB extends JpaRepository<PeminjamanRuanganModel, Integer> {
+public interface PeminjamanRuanganDB extends JpaRepository<PeminjamanRuanganModel, Long> {
     Optional<PeminjamanRuanganModel> findById(Integer id);
     Optional<PeminjamanRuanganModel> findByRuangan(RuanganModel ruanganModel);
 }
